refactor: AttendResultDto dto 패키지로 이동
